requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
sqlparse==0.4.3
python-dotenv==1.0.0
django_extensions==3.2.1
ipython==8.12.0
pytest-django>=4.5.2,<4.6
assertpy==1.1
phonenumbers>=8.13.11,<8.14
django-phonenumber-field==7.1.0
